Add S2S instruction-adherence bar chart
Collapse the WER accuracy chart and the new instruction chart into one
page-conditioned QualityBarSection over a shared QualityMetricBars chassis:
STT/TTS plot WER (lower is better, with the dataset toggle, dedicated
markers and click-to-compare), S2S plots instruction adherence (higher is
better, lean). BarTooltip is generalized with defaulted props so STT/TTS
are unchanged.
