+++ web/app/s2s/s2s-dashboard.tsx
@@ -10,9 +10,10 @@ import DashboardLayout from "@/components/layout/DashboardLayout";
 import { ChartSkeleton } from "@/components/dashboard/DashboardSkeleton";
 import { S2STopRow } from "./components/S2STopRow";
 
-// S2S has a single metric (V2V latency, no WER). The WER-based sections
-// (AccuracyBarSection, LatencyAccuracySection) are omitted; the model
-// comparison table renders WER-free (it hides the column when rows lack it).
+// S2S plots V2V latency plus instruction adherence (via QualityBarSection, the
+// shared quality-bar section — WER for STT/TTS, instruction here). The WER-only
+// LatencyAccuracySection/WerRadar are omitted; the model comparison table
+// renders WER-free (it hides the column when rows lack it).
 const TimelineChart = dynamic(
   () => import("@/components/visualizations/TimelineChart"),
   { ssr: false, loading: () => <ChartSkeleton /> }
@@ -20,6 +21,11 @@ const TimelineChart = dynamic(
 
 const BoxPlotSection = dynamic(
   () => import("@/components/dashboard/BoxPlotSection"),
+  { ssr: false, loading: () => <ChartSkeleton /> }
+);
+
+const QualityBarSection = dynamic(
+  () => import("@/components/dashboard/QualityBarSection"),
   { ssr: false, loading: () => <ChartSkeleton /> }
 );
 
@@ -36,6 +42,7 @@ export function S2SDashboard() {
           <S2STopRow />
           <TimelineChart />
           <BoxPlotSection />
+          <QualityBarSection />
           <ModelComparisonSection />
         </DashboardLayout>
       </SidebarMenuProvider>
