+++ web/app/stt/stt-dashboard.tsx
@@ -28,8 +28,8 @@ const LatencyAccuracySection = dynamic(
   { ssr: false, loading: () => <ChartSkeleton /> }
 );
 
-const AccuracyBarSection = dynamic(
-  () => import("@/components/dashboard/AccuracyBarSection"),
+const QualityBarSection = dynamic(
+  () => import("@/components/dashboard/QualityBarSection"),
   { ssr: false, loading: () => <ChartSkeleton /> }
 );
 
@@ -51,7 +51,7 @@ export function STTDashboard() {
           <KeyMetrics />
           <TimelineChart />
           <BoxPlotSection />
-          <AccuracyBarSection />
+          <QualityBarSection />
           <LatencyAccuracySection />
           <WerRadarSection />
           <ModelComparisonSection />
